app/requirements.txt
argcomplete==3.3.0
boto3==1.34.126
botocore==1.34.126
certifi==2024.6.2
cfn-flip==1.3.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
durationpy==0.6
Flask==3.0.3
Flask-Cors==4.0.1
hjson==3.1.0
idna==3.7
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.5
placebo==0.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.1
requests==2.32.3
s3transfer==0.10.1
six==1.16.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.66.4
troposphere==4.8.0
urllib3==2.2.1
Werkzeug==3.0.3
wheel==0.43.0
zappa==0.59.0